Removed lines associated with JDBC. Did formatting and removed unnecessary comments. Thank You.

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -3,7 +3,10 @@ package com.ust.pos.user;
 import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
+import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -23,7 +26,7 @@ public class UserController {
     private RoleService roleService;
 
     @GetMapping("/list")
-    public String home(Model model) {
+    public String list(Model model) {
         model.addAttribute("users", userService.findAll());
         return USER_LIST;
     }
@@ -47,7 +50,15 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(@RequestParam String username) {
-        userService.delete(username);
+        @Nullable Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/login";
+            }
+        }
         return REDIRECT_USER_LIST;
     }
 }
